requirements.txt
streamlit
google-generativeai
python-dotenv
gspread
oauth2client
pandas
pillow
